+++ index.js
@@ -23,87 +23,24 @@ const webhook = process.env.WEBHOOK;
 const errorwebhook = process.env.ERRORWEBHOOK;
 const dbltoken = process.env.DBLTOKEN;
 const developer = process.env.DEVELOPER;
+console.log(token);
 const client = new Discord.Client();
 const talkedRecently = new Set();
+
+var misc = require('./commands.js');
+var data = require('./data.js');
+var uwuifying = require('./uwuify.js');
+
 var Datastore = require("nedb"),
   blacklist = new Datastore({
     filename: ".data/datafile",
     autoload: true
   });
+
 const clean = text => {
   if (typeof text === "string") return text.replace(/`/g, "`" + String.fromCharCode(8203)).replace(/@/g, "@" + String.fromCharCode(8203));
   else return text;
 };
-const faces = ["(・`ω´・)", ";;w;;", "owo", "UwU", ">w<", "^w^", "✧w✧", "♥w♥", "(˘³˘)", "(。U⁄ ⁄ω⁄ ⁄ U。)", "(ᵘʷᵘ)", "(ᵕᴗ ᵕ⁎)", "uwU", "◔w◔", "⓪w⓪", "‿︵𝓇𝒶𝓌𝓇‿︵ ʘwʘ", "øwø", "ÓwÓ", "ÕwÕ", "@w@", "ᅌwᅌ", "ʘwʘ", "(✿◠‿◠)", "(●´ω｀●)", "(づ｡◕‿‿◕｡)づ", "≧◡≦", "(◡‿◡✿)", "(\*^ -^\*)", "(∪ ◡ ∪)", "(✿◠‿◠)", "╰(◡‿◡✿╰)", "(ﾉ◕ヮ◕)ﾉ\*:･ﾟ✧", "(￣ｰ￣)", "ヽ(゜∇゜)ノ", "(◕ω◕✿)", "(〃^∇^)ﾉ", "(\´｡• ᵕ •｡`)", "ヽ(>∀<☆)ノ", "ヽ(\*・ω・)ﾉ", "☆ ～('▽^人)", "(´ ω \`♡)", "(๑˃ᴗ˂)ﻭ", "( ´ ▽ \` ).｡ｏ♡", "╰(\*´︶`\*)╯♡", "ヽ(♡‿♡)ノ", "( ´ ∀ `)ノ～ ♡", "♡ ～('▽^人)", "( ´ ▽ \` ).｡ｏ♡", "Σ>―(〃°ω°〃)♡→", "(´,,•ω•,,)♡", "( ˘⌣˘)♡(˘⌣˘ )", "(„ಡωಡ„)", "(ノ\*°▽°\*)", "(｡･ω･｡)ﾉ♡", "(=^･ω･^=)", "╰(◡‿◡✿╰)", "(´･ω･\`)", "(=^-ω-^=)", "ヽ(=^･ω･^=)丿", "ʚ(\*´꒳\`\*)ɞ", "(´♡ω♡\`)", "★~(◡﹏◕✿)", "★~(◡ω◕✿)", "★~(◡﹏◡✿)", "★~(◠︿⊙✿)", "｡◕ ‿ ◕｡", "(◠︿◠✿)"];
-const statuses = ["with uwu faces", 'a game called "OwO what\'s this?!?"', "*notices shitpost* uwu what's this?", "@mention me to uwu-ify messages", "with s-senpai~~", "Coded by code monkeys at uwu headquawews", "Hewwo dewr~!", "Stay safe! uwu", "Stay home!", "Remember to wash your hands!"];
-
-function uwuify(str) {
-  str = str.replace(/god/g, "gawd");
-  str = str.replace(/God/g, "Gawd");
-  str = str.replace(/father/gi, "daddy");
-  str = str.replace(/papa/gi, "papi");
-  str = str.replace(/mom/g, "mommy");
-  str = str.replace(/mother/g, "mommy");
-  str = str.replace(/(?:r|l)/g, "w");
-  str = str.replace(/(?:R|L)/g, "W");
-  str = str.replace(/n([aeiou])/g, "ny$1");
-  str = str.replace(/N([aeiou])/g, "Ny$1");
-  str = str.replace(/N([AEIOU])/g, "NY$1");
-  str = str.replace(/ove/g, "uv");
-  str = str + "\~\~";
-  return str;
-}
-
-function errored(error, message) {
-  if (error.code == 50013) message.channel.send(":x: Oh no qwq! I don't have proper permissions to send you the content! Please make sure I have permissions to **Embed Links** in this server.").catch(() => {
-    message.author.send(":x: Oh no qwq! I don't have proper permissions to send you the content! Please make sure I have permissions to **Send Messages** in that server.").catch(() => {
-      axios.post(webhook, {
-        content: ":x: I tried sending an error DM to " + message.author.tag + ", but they have their DMs closed :|"
-      });
-    })
-  })
-  else {
-    const errorid = makeid(6);
-    axios.post(errorwebhook, {
-      content: "`" + errorid + "` - " + message.author.tag + " - ID: " + message.author.id + ":\n```" + error + "```"
-    });
-    message.channel.send(":warning: Something went wrong while executing your command. If you need more help, you can join my support server @ <https://discord.gg/eq6kwNJ> and give this code for error troubleshooting: `" + errorid + "`").catch(() => {
-      message.author.send(":warning: Something went wrong while executing your command. If you need more help, you can join my support server @ <https://discord.gg/eq6kwNJ> and give this code for error troubleshooting: `" + errorid + "`").catch(() => {
-        axios.post(errorwebhook, {
-          content: ":x: I tried sending a DM to " + message.author.tag + "about the error `" + errorid + "` but they have their DMs closed :|"
-        });
-      })
-    })
-  }
-};
-async function isBlacklisted(message) {
-  await blacklist.find({
-    userid: message.author.id
-  }, await function(err, result) {
-    if (result[0] != undefined) {
-      console.log("Found the user in the database");
-      return true;
-    }
-    if (err) {
-      console.log("Couldn't check if the user is blacklisted: " + err)
-      return false;
-    }
-    else {
-      console.log("Couldn't find the user in the blacklist");
-      return false;
-    }
-  });
-}
-
-function makeid(length) {
-  var result = "";
-  var characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  var charactersLength = characters.length;
-  for (var i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
 
 function statuschange() {
   client.user.setActivity(`${statuses[Math.floor(Math.random() * statuses.length)]} | @${
@@ -113,12 +50,14 @@ function statuschange() {
 process.on("unhandledRejection", error => {
   console.error("Unhandled promise rejection:", error);
 });
+
 client.on("ready", () => {
   axios.post(webhook, {
     content: "Logged in as " + client.user.tag + ". I can see " + client.users.cache.size + " users, in " + client.channels.cache.size + " channels of " + client.guilds.cache.size + " guilds."
   });
   console.log(`Logged in as ${client.user.tag}. I can see ${client.users.cache.size} users, in ${client.channels.cache.size} channels of ${client.guilds.cache.size} guilds.`);
   setInterval(statuschange, 120000);
+  console.log(misc)
 });
 client.on("guildCreate", guild => {
   axios.post(webhook, {
@@ -132,11 +71,28 @@ client.on("guildDelete", guild => {
   });
   console.log(`I have been removed from: ${guild.name} (id: ${guild.id})`);
 });
-client.on("message", async message => {
+client.on("message", message => {
   if (message.author.bot || message.channel.type != "text") return;
   const messagebutstring = message.content;
   if (messagebutstring.startsWith("<@!" + client.user.id + ">") || messagebutstring.startsWith("<@" + client.user.id + ">")) {
-    if (isBlacklisted(message) == true) {
+    var isBlacklisted;
+    blacklist.find({
+     userid: message.author.id
+    }, function(err, result) {
+      if (result[0] != undefined) {
+          console.log("Found the user in the database");
+          isBlacklisted = true;
+        }
+      else if (err) {
+          console.log("Couldn't check if the user is blacklisted: " + err)
+          isBlacklisted = false;
+      } else {
+          console.log("Couldn't find the user in the blacklist");
+          isBlacklisted = false;
+        }
+    });
+  };
+    if (isBlacklisted == true) {
       console.log("Responding with the blacklist message");
       message.react("🚫");
       message.author.send("🚫 You have been blacklisted from using the bot for not following our Terms of Service. If you would like to appeal, please join our server @ <https://discord.gg/eq6kwNJ> and head over to #support to appeal.\n\nYou can find our Terms of Service here: https://github.com/TheOnlyGhostwolf/uwutranslator/wiki/Terms-of-Service");
@@ -160,120 +116,30 @@ client.on("message", async message => {
       if (command == "" || command == " ") {
         message.channel.send("Hewwo <@" + message.author.id + ">! (^w^)/\n\nI'm **" + client.user.username + "**, I uwu-ify messages. If you want to check how to use me, use **<@!" + client.user.id + "> --help** command :3").catch(error => errored(error, message));
       } else if (command === "--ping") {
-        const m = await message.channel.send("Ping?").catch(error => errored(error, message));
-        m.edit(`Pong! Latency is ${m.createdTimestamp -
-            message.createdTimestamp}ms. API Latency is ${Math.round(
-            client.ping
-          )}ms`).catch(error => errored(error, message));
+        misc.commands.ping(message);
+        console.log("LEt's fucking do it");
       } else if (command === "--shutdown") {
-        let isBotOwner = message.author.id == developer;
-        if (!isBotOwner) {
-          message.channel.send(":warning: Only the bot developer can use this command");
-          return;
-        }
-        message.channel.send("I-I don't feel so good... qwq | Shutting down...").then(m => {
-          client.destroy();
-          process.exit(1);
-        });
+        misc.commands.shutdown(message, developer);
       } else if (command == "--blacklist") {
-        let isBotOwner = message.author.id == developer;
-        if (!isBotOwner) {
-          message.channel.send(":warning: Only the bot developer can use this command");
-          return;
-        }
-        blacklist.find({
-          userid: args
-        }, function(err, result) {
-          if (result[0] == undefined) {
-            blacklist.insert({
-              userid: args
-            }, function(err, userAdded) {
-              if (err) {
-                console.log("There's a problem with the database: ", err);
-                message.channel.send(
-                  ":x: There was an error while trying to add the user to the blacklist. Please try again later."
-                );
-              } else if (userAdded) {
-                console.log(
-                  "New user with ID " + args + " inserted to the blacklist"
-                );
-                message.channel.send(
-                  "New user with ID " + args + " inserted to the blacklist"
-                );
-              }
-            });
-          } else {
-            blacklist.remove({
-              userid: args
-            }, {
-              multi: true
-            }, function(err) {
-              if (err) {
-                console.log("There's a problem with the database: ", err);
-                message.channel.send(
-                  ":x: There was an error while trying to remove the user from the blacklist. Please try again later."
-                );
-              } else {
-                console.log("User " + args + " removed from the blacklist");
-                message.channel.send(
-                  "User " + args + " removed from the blacklist"
-                );
-              }
-            });
-          }
-        })
+        misc.commands.eval(message, developer);
       } else if (command === "--eval") {
-        let isBotOwner = message.author.id == developer;
-        if (!isBotOwner) {
-          message.channel.send(":warning: Only the bot developer can use this command");
-          return;
-        }
-        try {
-          const code = args.join(" ");
-          let evaled = eval(code);
-          if (typeof evaled !== "string") evaled = require("util").inspect(evaled);
-          message.channel.send(clean(evaled), {
-            code: "xl"
-          }).catch(error => errored(error, message));
-        } catch (err) {
-          message.channel.send(`\`ERROR\` \`\`\`xl\n${clean(err)}\n\`\`\``);
-        }
+        misc.commands.blacklist(message, developer);
       } else if (command === "--help") {
-        var helpembed = new Discord.MessageEmbed({
-            title: "**Hewwo! I'm uwutranslator!**",
-            description: "I uwu-ify messages. @mention me and type any text for me to translate! >wO\n\nExample: **@uwutranslator Hello world! I am alive!**\n"
-          })
-          .setFooter(message.author.displayAvatarURL, message.author.username + "#" + message.author.discriminator)
-          .setThumbnail("https://media.giphy.com/media/VUC9YdLSnKuJy/giphy.gif")
-          .addField("Developer", "[Ghostwolf#6735](https://ghostwolf.me)", true)
-          .addField("Add me to your server!", "[Click here](https://discordapp.com/oauth2/authorize?client_id=635507578008240165&permissions=84992&scope=bot)", true)
-          .addField("Special thanks", "Tea, Elias, Dragonic, Pretzel", true)
-          .addField("Website", "[Click here](https://uwutranslator.ghostwolf.me)", true)
-          .addField("Support the developer!", "[Buy me a coffee!](https://ko-fi.com/ghostwolf)", true)
-          .addField("Vote for me on DBL!", "[Click here](https://top.gg/bot/635507578008240165/vote)", true)
-          .addField("Legal Mumbo Jumbo", "[Terms of Service](https://github.com/TheOnlyGhostwolf/uwutranslator/wiki/Terms-of-Service) | [Privacy Policy](https://github.com/TheOnlyGhostwolf/uwutranslator/wiki/Privacy-Policy)", true)
-          .setColor(16761576)
-          .setTimestamp(message.createdAt)
-          .setFooter("Requested by " + message.author.username + "#" + message.author.discriminator, message.author.avatarURL)
-        message.channel.send(helpembed).catch(error => errored(error, message));
+        misc.commands.help(message);
       } else if (str.includes("discord.gg") || str.includes("discordapp.com/invite")) {
         message.reply("don't send invite links using me >:(").catch(error => errored(error, message));
       } else {
-        if (str.slice(-1) == " ") str = str.substring(0, str.length - 1);
-        var uwuifiedstr = uwuify(str);
-        var firstletter = uwuifiedstr.substring(0, 1);
-        uwuifiedstr = firstletter + "-" + uwuifiedstr + " " + faces[Math.floor(Math.random() * faces.length)] + " ";
-        let uwuembed = new Discord.MessageEmbed({
-          description: uwuifiedstr
-        });
-        uwuembed.setColor(16761576);
-        uwuembed.setFooter("Requested by " + message.author.tag + " | @mention me to uwu-ify messages", message.author.avatarURL());
-        message.channel.send(uwuembed).catch(error => errored(error, message));
+        uwuifying.uwuify.custom(message);
       }
     }
   }
-});
+);
 client.login(token);
+
+module.exports.commands = misc;
+module.exports.uwuify = uwuifying;
+module.exports.data = data;
+
 /*
 // REMOVE THIS IF YOUR BOT ISN'T LISTED ON TOP.GG
 const DBL = require("dblapi.js");
