+++ package.json
@@ -13,6 +13,8 @@
   },
   "author": "Paulius Gečas",
   "license": "GPL-3.0-only",
+  "engine": ">=13.2",
+  "engineStrict": true,
   "bugs": {
     "url": "https://github.com/TheOnlyGhostwolf/uwutranslator/issues"
   },
